"Enemy parent Class, AntEnemy HornetEnemy children"
