--- mejiavea-experiment-file.py
@@ -1,9 +1,0 @@
-wave_1_10 = ["ant" for _ in range(49)]
-wave_1_10 += ["hornet"]
-print(wave_1_10)
-
-wave_11_20 = ["ant" for _ in range(20)]
-wave_11_20 += ["hornet" for _ in range(5)]
-wave_11_20 += ["ant" for _ in range(15)]
-wave_11_20 += ["hornet" for _ in range(10)]
-print(wave_11_20)
